+++ JsonToMidiConverter/Models/Song/BrushStroke.cs
@@ -5,8 +5,8 @@ namespace JsonToMidiConverter.Models.Song;
 public class BrushStroke
 {
     public Direction Direction { get; set; }
-    public double Duration { get; set; }
-    public double Shift { get; set; }
+    public short Duration { get; set; }
+    public float Shift { get; set; } // theres a record with value "28.000001" maybe we can just normalize the data
 
     public BrushStroke Clone() => new()
     {
