+++ JsonToMidiConverter/Models/Song/ChordStroke.cs
@@ -6,6 +6,6 @@ namespace JsonToMidiConverter.Models.Song;
 public sealed class ChordStroke : ISerializable
 {
     public StrokeTechnique Technique { get; set; } = StrokeTechnique.None;
-    public double Duration { get; set; }
-    public double StartTimeOffset { get; set; }
+    public short Duration { get; set; } // 0 - 6
+    public float StartTimeOffset { get; set; } // 0 - 100
 }
