+++ JsonToMidiConverter/Models/Song/Marker.cs
@@ -3,7 +3,7 @@
 public sealed class Marker : ISerializable
 {
     public string Text { get; set; } = string.Empty;
-    public double Width { get; set; }
+    public short Width { get; set; } // 0 - 565
 
     public Marker Clone() => new()
     {
